+++ src/Opaleye/Internal/PGTypesExternal.hs
@@ -13,15 +13,16 @@ import qualified Opaleye.Internal.PGTypes               as IPT
 import qualified Opaleye.Internal.HaskellDB.PrimQuery   as HPQ
 import qualified Opaleye.Internal.HaskellDB.Sql.Default as HSD
 
-import qualified Data.Aeson                             as Ae
-import qualified Data.ByteString                        as SByteString
-import qualified Data.ByteString.Lazy                   as LByteString
-import qualified Data.CaseInsensitive                   as CI
-import           Data.Scientific                        as Sci
-import qualified Data.Text                              as SText
-import qualified Data.Text.Lazy                         as LText
-import qualified Data.Time                              as Time
-import qualified Data.UUID                              as UUID
+import qualified Data.CaseInsensitive as CI
+import qualified Data.Aeson as Ae
+import qualified Data.Text as SText
+import qualified Data.Text.Lazy as LText
+import qualified Data.ByteString as SByteString
+import qualified Data.ByteString.Lazy as LByteString
+import           Data.Scientific as Sci
+import qualified Data.Time.Compat as Time
+import qualified Data.Time.Format.ISO8601.Compat as Time.Format.ISO8601
+import qualified Data.UUID as UUID
 
 import           Data.Int                               (Int64)
 
@@ -50,6 +51,9 @@ instance C.SqlIntegral SqlInt8
 instance C.SqlString SqlText where
   sqlFromString = pgString
 
+instance C.SqlString SqlVarcharN where
+  sqlFromString = sqlStringVarcharN
+
 instance C.SqlString SqlCitext where
   sqlFromString = pgCiLazyText . CI.mk . LText.pack
 
@@ -70,6 +74,15 @@ pgStrictText = IPT.literalColumn . HPQ.StringLit . SText.unpack
 pgLazyText :: LText.Text -> Column PGText
 pgLazyText = IPT.literalColumn . HPQ.StringLit . LText.unpack
 
+sqlStringVarcharN :: String -> Column SqlVarcharN
+sqlStringVarcharN = IPT.literalColumn . HPQ.StringLit
+
+sqlStrictTextVarcharN :: SText.Text -> Column SqlVarcharN
+sqlStrictTextVarcharN = IPT.literalColumn . HPQ.StringLit . SText.unpack
+
+sqlLazyTextVarcharN :: LText.Text -> Column SqlVarcharN
+sqlLazyTextVarcharN = IPT.literalColumn . HPQ.StringLit . LText.unpack
+
 pgNumeric :: Sci.Scientific -> Column PGNumeric
 pgNumeric = IPT.literalColumn . HPQ.NumericLit
 
@@ -89,23 +102,25 @@ pgUUID :: UUID.UUID -> Column PGUuid
 pgUUID = IPT.literalColumn . HPQ.StringLit . UUID.toString
 
 pgDay :: Time.Day -> Column PGDate
-pgDay = IPT.unsafePgFormatTime "date" "'%0Y-%m-%d'"
+pgDay = IPT.unsafePgFormatTime "date"
 
 pgUTCTime :: Time.UTCTime -> Column PGTimestamptz
-pgUTCTime = IPT.unsafePgFormatTime "timestamptz" "'%0Y-%m-%dT%T%QZ'"
+pgUTCTime = IPT.unsafePgFormatTime "timestamptz"
 
 pgLocalTime :: Time.LocalTime -> Column PGTimestamp
-pgLocalTime = IPT.unsafePgFormatTime "timestamp" "'%0Y-%m-%dT%T%Q'"
+pgLocalTime = IPT.unsafePgFormatTime "timestamp"
 
 pgZonedTime :: Time.ZonedTime -> Column PGTimestamptz
-pgZonedTime = IPT.unsafePgFormatTime "timestamptz" "'%0Y-%m-%dT%T%Q%z'"
+pgZonedTime = IPT.unsafePgFormatTime "timestamptz"
 
 pgTimeOfDay :: Time.TimeOfDay -> Column PGTime
-pgTimeOfDay = IPT.unsafePgFormatTime "time" "'%T%Q'"
+pgTimeOfDay = IPT.unsafePgFormatTime "time"
 
 -- "We recommend not using the type time with time zone"
 -- http://www.postgresql.org/docs/8.3/static/datatype-datetime.html
 
+sqlInterval :: Time.CalendarDiffTime -> Column PGInterval
+sqlInterval = IPT.unsafePgFormatTime "interval"
 
 pgCiStrictText :: CI.CI SText.Text -> Column PGCitext
 pgCiStrictText = IPT.literalColumn . HPQ.StringLit . SText.unpack . CI.original
@@ -179,10 +194,14 @@ instance IsSqlType SqlInt4 where
   showSqlType _ = "integer"
 instance IsSqlType SqlInt2 where
   showSqlType _ = "smallint"
+instance IsSqlType SqlInterval where
+  showSqlType _ = "interval"
 instance IsSqlType SqlNumeric where
   showSqlType _ = "numeric"
 instance IsSqlType SqlText where
   showSqlType _ = "text"
+instance IsSqlType SqlVarcharN where
+  showSqlType _ = "varchar"
 instance IsSqlType SqlTime where
   showSqlType _ = "time"
 instance IsSqlType SqlTimestamp where
@@ -234,10 +253,29 @@ data SqlFloat8
 data SqlInt8
 data SqlInt4
 data SqlInt2
+-- | Requires you to configure @intervalstyle@ as @iso_8601@.
+--
+-- You can configure @intervalstyle@ on every connection with a @SET@ command,
+-- but for better performance you may want to configure it permanently in the
+-- file found with @SHOW config_file;@.
+data SqlInterval
 data SqlNumeric
 data SqlText
+-- | @VARCHAR(n)@ for any @n@.  Opaleye does not do anything to check
+-- that the @n@ you choose is correctly adhered to!
+data SqlVarcharN
 data SqlTime
 data SqlTimestamp
+-- | Be careful if you use Haskell's `Time.ZonedTime` with
+-- @SqlTimestamptz@. A Postgres @timestamptz@ does not actually
+-- contain any time zone.  It is just a UTC time that is automatically
+-- converted to or from local time on certain occasions, according to
+-- the [timezone setting of the
+-- server](https://www.postgresql.org/docs/9.1/runtime-config-client.html#GUC-TIMEZONE).
+-- Therefore, although when you roundtrip an input 'Time.ZonedTime' to
+-- obtain an output 'Time.ZonedTime' they each refer to the same
+-- instant in time, the time zone attached to the output will not
+-- necessarily the same as the time zone attached to the input.
 data SqlTimestamptz
 data SqlUuid
 data SqlCitext
@@ -254,6 +292,7 @@ type PGFloat8 = SqlFloat8
 type PGInt8 = SqlInt8
 type PGInt4 = SqlInt4
 type PGInt2 = SqlInt2
+type PGInterval = SqlInterval
 type PGNumeric = SqlNumeric
 type PGText = SqlText
 type PGTime = SqlTime