+++ src/Opaleye/Internal/PrimQuery.hs
@@ -21,6 +21,8 @@ data BinOp = Except
 
 data JoinType = LeftJoin | RightJoin | FullJoin deriving Show
 
+data SemijoinType = Semi | Anti deriving Show
+
 data TableIdentifier = TableIdentifier
   { tiSchemaName :: Maybe String
   , tiTableName  :: String
@@ -55,6 +57,8 @@ data PrimQuery' a = Unit
                   | Empty     a
                   | BaseTable TableIdentifier (Bindings HPQ.PrimExpr)
                   | Product   (NEL.NonEmpty (Lateral, PrimQuery' a)) [HPQ.PrimExpr]
+                  -- | The subqueries to take the product of and the
+                  --   restrictions to apply
                   | Aggregate (Bindings (Maybe (HPQ.AggrOp,
                                                 [HPQ.OrderExpr],
                                                 HPQ.AggrDistinct),
@@ -72,11 +76,10 @@ data PrimQuery' a = Unit
                   | Limit     LimitOp (PrimQuery' a)
                   | Join      JoinType
                               HPQ.PrimExpr
-                              (Bindings HPQ.PrimExpr)
-                              (Bindings HPQ.PrimExpr)
-                              (PrimQuery' a)
-                              (PrimQuery' a)
-                  | Exists    Bool (PrimQuery' a) (PrimQuery' a)
+                              (Lateral, PrimQuery' a)
+                              (Lateral, PrimQuery' a)
+                  | Semijoin  SemijoinType (PrimQuery' a) (PrimQuery' a)
+                  | Exists    Symbol (PrimQuery' a)
                   | Values    [Symbol] (NEL.NonEmpty [HPQ.PrimExpr])
                   | Binary    BinOp
                               (PrimQuery' a, PrimQuery' a)
@@ -112,12 +115,11 @@ data PrimQueryFold' a p = PrimQueryFold
   , limit             :: LimitOp -> p -> p
   , join              :: JoinType
                       -> HPQ.PrimExpr
-                      -> Bindings HPQ.PrimExpr
-                      -> Bindings HPQ.PrimExpr
+                      -> (Lateral, p)
+                      -> (Lateral, p)
                       -> p
-                      -> p
-                      -> p
-  , existsf           :: Bool -> p -> p -> p
+  , semijoin          :: SemijoinType -> p -> p -> p
+  , exists            :: Symbol -> p -> p
   , values            :: [Symbol] -> NEL.NonEmpty [HPQ.PrimExpr] -> p
   , binary            :: BinOp
                       -> (p, p)
@@ -140,11 +142,12 @@ primQueryFoldDefault = PrimQueryFold
   , distinctOnOrderBy = DistinctOnOrderBy
   , limit             = Limit
   , join              = Join
+  , semijoin          = Semijoin
   , values            = Values
   , binary            = Binary
   , label             = Label
   , relExpr           = RelExpr
-  , existsf           = Exists
+  , exists            = Exists
   , rebind            = Rebind
   , forUpdate         = ForUpdate
   }
@@ -159,27 +162,22 @@ foldPrimQuery f = fix fold
           Aggregate aggrs q           -> aggregate         f aggrs (self q)
           DistinctOnOrderBy dxs oxs q -> distinctOnOrderBy f dxs oxs (self q)
           Limit op q                  -> limit             f op (self q)
-          Join j cond pe1 pe2 q1 q2   -> join              f j cond pe1 pe2 (self q1) (self q2)
+          Join j cond q1 q2           -> join              f j cond (fmap self q1) (fmap self q2)
+          Semijoin j q1 q2            -> semijoin          f j (self q1) (self q2)
           Values ss pes               -> values            f ss pes
           Binary binop (q1, q2)       -> binary            f binop (self q1, self q2)
           Label l pq                  -> label             f l (self pq)
           RelExpr pe syms             -> relExpr           f pe syms
-          Exists b q1 q2              -> existsf           f b (self q1) (self q2)
+          Exists s q                  -> exists            f s (self q)
           Rebind star pes q           -> rebind            f star pes (self q)
           ForUpdate q                 -> forUpdate         f (self q)
         fix g = let x = g x in x
 
-times :: PrimQuery -> PrimQuery -> PrimQuery
-times q q' = Product (pure q NEL.:| [pure q']) []
+times :: Lateral -> PrimQuery -> PrimQuery -> PrimQuery
+times lat q q' = Product (pure q NEL.:| [(lat, q')]) []
 
 restrict :: HPQ.PrimExpr -> PrimQuery -> PrimQuery
 restrict cond primQ = Product (return (pure primQ)) [cond]
-
-exists :: PrimQuery -> PrimQuery -> PrimQuery
-exists = Exists True
-
-notExists :: PrimQuery -> PrimQuery -> PrimQuery
-notExists = Exists False
 
 isUnit :: PrimQuery' a -> Bool
 isUnit Unit = True