+++ src/Opaleye/SqlTypes.hs
@@ -2,29 +2,90 @@
 -- those types.  To create fields you may find it more convenient to use
 -- "Opaleye.ToFields" instead.
 
-module Opaleye.SqlTypes (module Opaleye.SqlTypes,
-                         P.IsSqlType,
-                         P.IsRangeType,
-                         SqlBool,
-                         SqlDate,
-                         SqlFloat4,
-                         SqlFloat8,
-                         SqlInt8,
-                         SqlInt4,
-                         SqlInt2,
-                         SqlNumeric,
-                         SqlText,
-                         SqlTime,
-                         SqlTimestamp,
-                         SqlTimestamptz,
-                         SqlUuid,
-                         SqlCitext,
-                         SqlArray,
-                         SqlBytea,
-                         SqlJson,
-                         SqlJsonb,
-                         SqlRange,
-                        ) where
+module Opaleye.SqlTypes (
+  -- * Numeric
+  -- ** Creating values
+  sqlInt4,
+  sqlDouble,
+  sqlInt8,
+  sqlNumeric,
+  -- ** Types
+  SqlInt4,
+  SqlFloat8,
+  SqlNumeric,
+  SqlInt8,
+  SqlInt2,
+  SqlFloat4,
+  -- * Date and time
+  -- ** Creating values
+  sqlDay,
+  sqlUTCTime,
+  sqlLocalTime,
+  sqlZonedTime,
+  sqlTimeOfDay,
+  P.sqlInterval,
+  -- ** Types
+  SqlDate,
+  SqlTime,
+  SqlTimestamp,
+  SqlTimestamptz,
+  SqlInterval,
+  -- * JSON
+  -- ** Creating values
+  sqlJSON,
+  sqlStrictJSON,
+  sqlLazyJSON,
+  sqlValueJSON,
+  -- ** Types
+  SqlJson,
+  -- * JSONB
+  -- ** Creating values
+  sqlJSONB,
+  sqlStrictJSONB,
+  sqlLazyJSONB,
+  sqlValueJSONB,
+  -- ** Types
+  SqlJsonb,
+  -- * Text
+  -- ** Creating values
+  sqlString,
+  sqlStrictText,
+  sqlLazyText,
+  P.sqlStringVarcharN,
+  P.sqlStrictTextVarcharN,
+  P.sqlLazyTextVarcharN,
+  sqlCiStrictText,
+  sqlCiLazyText,
+  -- ** Types
+  SqlText,
+  SqlVarcharN,
+  SqlCitext,
+  -- * Array
+  -- ** Creating values
+  sqlArray,
+  -- ** Types
+  SqlArray,
+  -- * Range
+  -- ** Creating values
+  sqlRange,
+  -- ** Types
+  SqlRange,
+  P.IsRangeType,
+  -- * Other
+  -- ** Creating values
+  sqlBool,
+  sqlUUID,
+  sqlLazyByteString,
+  sqlStrictByteString,
+  -- ** Types
+  SqlBool,
+  SqlUuid,
+  SqlBytea,
+  -- * @IsSqlType@
+  P.IsSqlType,
+  -- * Entire module
+  module Opaleye.SqlTypes,
+  ) where
 
 import qualified Opaleye.Field   as F
 import qualified Opaleye.Internal.PGTypesExternal as P
@@ -38,9 +99,11 @@ import           Opaleye.Internal.PGTypesExternal (SqlBool,
                                                    SqlInt2,
                                                    SqlNumeric,
                                                    SqlText,
+                                                   SqlVarcharN,
                                                    SqlTime,
                                                    SqlTimestamp,
                                                    SqlTimestamptz,
+                                                   SqlInterval,
                                                    SqlUuid,
                                                    SqlCitext,
                                                    SqlArray,
@@ -57,7 +120,7 @@ import           Data.Int (Int64)
 import           Data.Scientific as Sci
 import qualified Data.Text as SText
 import qualified Data.Text.Lazy as LText
-import qualified Data.Time as Time
+import qualified Data.Time.Compat as Time
 import qualified Data.UUID as UUID
 
 import qualified Database.PostgreSQL.Simple.Range as R
@@ -114,7 +177,6 @@ sqlTimeOfDay = P.pgTimeOfDay
 
 -- "We recommend not using the type time with time zone"
 -- http://www.postgresql.org/docs/8.3/static/datatype-datetime.html
-
 
 sqlCiStrictText :: CI.CI SText.Text -> F.Field SqlCitext
 sqlCiStrictText = P.pgCiStrictText